Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonify==0.5
MarkupSafe==3.0.2
mongoengine==0.29.1
numpy==2.2.2
openai==1.61.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
